Fixed errors from w3 validator
No real changes, just made everything pretty.  Only error we really get is that the div inside our nav button for mobile is causing some sort of contextual thing on every page.  We'll just have to tell Mr. E that's what we have to do according to w3 schools. Also on index it says date is not supported on all browsers and that the link for the weather widget is broken, but it's not.